requirements.txt
pytest==7.2.0
PyMySQL==1.0.2
PyYAML==5.3.1
allure-pytest==2.11.1
requests==2.28.1
